+++ src/main/scala/org/calinburloiu/music/tuning/Tuning.scala
@@ -16,45 +16,32 @@
 
 package org.calinburloiu.music.tuning
 
-import com.google.common.base.Preconditions._
-
 /**
- * Describes the tuning of keyed instrument, typically with a piano keyboard, by specifying a deviation in cents
+ * Describes the tuning of a keyed instrument, typically with a piano keyboard, by specifying a deviation in cents
  * for each 12-EDO key.
  *
- * A typically piano tuning might be specified only for each of the 12 pitch classes: C, C#\Db, ..., B. A tunings
+ * A typical piano tuning might be specified only for each of the 12 pitch classes: C, C#\Db, ..., B. A tuning
  * might use a different value for each 88 piano keys.
- *
- * @param name       a human-friendly name for the tuning
- * @param deviations deviation in cents for each key
+ * @tparam U
  */
-case class Tuning(name: String,
-                  override val deviations: Seq[Double]) extends TuningBase[Double] {
+trait Tuning[U] extends Iterable[U] {
+  require(deviations.nonEmpty, "Expecting a non-empty list of deviations")
 
-  /** Returns the deviation in cents for a particular key 0-based index. */
-  def apply(index: Int): Double = {
-    checkElementIndex(index, size)
-    deviations(index)
-  }
+  def name: String
+
+  def deviations: Seq[U]
+
+  def apply(index: Int): U
 
   /**
    * @return the size of the tuning
    */
   override def size: Int = deviations.size
 
-  override def iterator: Iterator[Double] = deviations.iterator
+  override def iterator: Iterator[U] = deviations.iterator
 
   override def toString: String = {
     val superString = super.toString
     s"$name = $superString"
   }
-}
-
-object Tuning {
-
-  /** The tuning for a 12-tone equal temperament, which has 0 cents deviation for each of the 12-keys. */
-  val Edo12: Tuning = Tuning("Equal Temperament", Seq.fill(12)(0.0))
-
-  def apply(name: String, headDeviation: Double, tailDeviations: Double*): Tuning =
-    Tuning(name, headDeviation +: tailDeviations)
 }
