--- src/main/scala/org/calinburloiu/music/tuning/TuningBase.scala
@@ -1,27 +1,0 @@
-/*
- * Copyright 2020 Calin-Andrei Burloiu
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package org.calinburloiu.music.tuning
-
-import com.google.common.base.Preconditions._
-
-trait TuningBase[U] extends Iterable[U] {
-  checkArgument(deviations.nonEmpty, "Expecting a non-empty list of deviations".asInstanceOf[Object])
-
-  def apply(index: Int): U
-
-  def deviations: Seq[U]
-}
