+++ src/test/scala/org/calinburloiu/music/microtuner/MonophonicPitchBendTunerTest.scala
@@ -17,7 +17,7 @@
 package org.calinburloiu.music.microtuner
 
 import org.calinburloiu.music.microtuner.midi.{MidiNote, MidiProcessorFixture, PitchBendSensitivity, Rpn, ScCcMidiMessage, ScNoteOffMidiMessage, ScNoteOnMidiMessage, ScPitchBendMidiMessage}
-import org.calinburloiu.music.tuning.Tuning
+import org.calinburloiu.music.tuning.OctaveTuning
 import org.scalactic.{Equality, TolerantNumerics}
 import org.scalatest.Inside
 import org.scalatest.flatspec.AnyFlatSpec
@@ -31,7 +31,7 @@ class MonophonicPitchBendTunerTest extends AnyFlatSpec with Matchers with Inside
   private val pitchBendSensitivity = PitchBendSensitivity(1)
 
   //@formatter:off
-  private val customTuning = Tuning(
+  private val customTuning = OctaveTuning(
     "major-ish in 72-EDO",
     0.0,    // C
     16.67,  // Db (~16/15 from C)
@@ -47,7 +47,7 @@ class MonophonicPitchBendTunerTest extends AnyFlatSpec with Matchers with Inside
     -16.67  // B
   )
   //@formatter:on
-  private val customTuning2 = Tuning("custom2", -45.0, -34.0, -23.0, -12.0, -1, 2, 13, 24, 35, 46, 17, 34)
+  private val customTuning2 = OctaveTuning("custom2", -45.0, -34.0, -23.0, -12.0, -1, 2, 13, 24, 35, 46, 17, 34)
 
   val Seq(noteC4, noteDFlat4, noteD4, noteDSharp4, noteE4, noteF4, noteFSharp4, noteG4,
     noteAb4, noteA4, noteBb4, noteB4) = 60 until 72
