+++ src/test/scala/org/calinburloiu/music/microtuner/midi/MidiTuningMessageGeneratorTest.scala
@@ -16,7 +16,7 @@
 
 package org.calinburloiu.music.microtuner.midi
 
-import org.calinburloiu.music.tuning.Tuning
+import org.calinburloiu.music.tuning.OctaveTuning
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 
@@ -36,8 +36,8 @@ class MidiTuningMessageGeneratorTest extends AnyFunSuite with Matchers {
 
   // TODO @calinburloiu Ignoring. It was using deprecated Java API.
   ignore("MidiTuningMessageGenerator.NonRealTime1BOctave") {
-    val tuning = Tuning("t", 0.1, 0.9, 2.4, 3.0, 4.2, 4.6, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0)
-    val expectedTuning = Tuning("t", 0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0)
+    val tuning = OctaveTuning("t", 0.1, 0.9, 2.4, 3.0, 4.2, 4.6, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0)
+    val expectedTuning = OctaveTuning("t", 0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0)
     val messageGenerator = MtsTuningFormat.NonRealTime1BOctave.messageGenerator
     messageGenerator should be theSameInstanceAs MidiTuningMessageGenerator.NonRealTime1BOctave
